* develop:
  [POTATO-20] 롬복 대체가능한거 수정 및 학과 추가 (#97)
  [POTATO-19] 관리자 모듈 세션 정보 memberId -> AdminId로 네이밍 수정 (#96)
  [POTATO-18] 게시물 정보(제목, 내용, 사진) 및 백업 정보 Value Object로 리팩토링 (#95)
  [POTATO-18] http client
  [POTATO-16] 게시물 도메인 설계 수정
  [POTATO-15] 블루 서버 켜질때까지 기다렸다가 그린 서버 종료하도록 수정
  [POTATO-14] 게시물 테이블에 counts 필드 추가
  [POTATO-14] 게시물 댓글 수정 (#94)
  Feat/like (#93)
  UPDATE README.md
  [POTATO-13] JDBC 세션을 위한 테이블, 인덱스 제거
  [POTATO-12] Api, Admin 서버 로그 도커 볼륨 마운트 설정.
  v0.3.2

+++ deploy/run.sh
@@ -10,16 +10,43 @@ if [ -z "$EXIST_BLUE" ]; then
     echo "Blue Up"
     docker-compose -p ${DOCKER_APP_NAME}-blue -f docker-compose.blue.yml up -d --build
 
-    sleep 60
+    i=0
+    while [ $i -le 30 ]
+    do
+        STATUS=$(curl -s -o /dev/null -w "%{http_code}" http://localhost:8081/ping)
+        if [ $STATUS -eq 200 ]; then
+            echo "Server is Running.....!"
 
-    echo "Green Down"
-    docker-compose -p ${DOCKER_APP_NAME}-green -f docker-compose.green.yml down
+            docker-compose -p ${DOCKER_APP_NAME}-green -f docker-compose.green.yml down
+            echo "Green Down"
+
+            break
+        else
+            echo "Waiting........."
+            sleep 10
+            i=$(($i+1))
+        fi
+    done
 else
     echo "Green Up"
     docker-compose -p ${DOCKER_APP_NAME}-green -f docker-compose.green.yml up -d --build
 
-    sleep 60
+    i=0
+    while [ $i -le 30 ]
+    do
+        STATUS=$(curl -s -o /dev/null -w "%{http_code}" http://localhost:8082/ping)
+        if [ $STATUS -eq 200 ]; then
+            echo "Server is Running.....!"
 
-    echo "Blue Down"
-    docker-compose -p ${DOCKER_APP_NAME}-blue -f docker-compose.blue.yml down
+            docker-compose -p ${DOCKER_APP_NAME}-blue -f docker-compose.blue.yml down
+            echo "Blue Down"
+
+            break
+        else
+            echo "Waiting........."
+            sleep 10
+            i=$(($i+1))
+        fi
+    done
 fi
+
